+++ lib/ts/recipe/thirdparty/recipeImplementation.ts
@@ -3,7 +3,6 @@ import { User } from "../authRecipeModule/types";
 import { NormalisedAppInfo } from "../../types";
 import Querier from "../../querier";
 import { appendQueryParamsToURL, getWindowOrThrow, getQueryParams } from "../../utils";
-import Session from "../session/recipe";
 
 export default class RecipeImplementation implements RecipeInterface {
     querier: Querier;
@@ -80,10 +79,6 @@ export default class RecipeImplementation implements RecipeInterface {
                     ...context,
                     action: "SIGN_IN",
                 });
-            },
-            async (context) => {
-                Session.getInstanceOrThrow().recipeImpl.saveSessionFromResponse(context);
-                return context.response;
             }
         );